parallel_port/requirements.txt
asyncio== 0.4.1
websockets
pyparallel